+++ rusmpp-extra/src/concatenation/owned/multipart/mod.rs
@@ -1,2 +1,5 @@
 mod submit_sm;
-pub use submit_sm::{SubmitSmMultipartBuilder, SubmitSmMultipartExt};
+pub use submit_sm::SubmitSmMultipartExt;
+
+mod udh;
+pub use udh::SubmitSmMultipartBuilder;
